# perf(storage): multi-file torrent IO with precomputation and binary search