1. Precomputing file paths and cumulative offsets in OpenTorrent to eliminate repeated path construction and linear scans during IO.
2. binary search in ReadAt and WriteAt to locate the starting file for a given offset, reducing time complexity from O(n) to O(log n)
3. Pre-creating all directories upfront in OpenTorrent to avoid redundant syscalls during writes.

**Before**
```
file_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/storage
cpu: Apple M1 Pro
BenchmarkFileStorage/WriteAtLastByte-10                   55      23473835 ns/op    67110972 B/op    1048590 allocs/op
PASS
ok      github.com/james-lawrence/torrent/storage   1.706s
```

**After**
```
file_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/storage
cpu: Apple M1 Pro
BenchmarkFileStorage/WriteAtLastByte-10                65466         18335 ns/op         248 B/op          3 allocs/op
PASS
ok      github.com/james-lawrence/torrent/storage   4.733s
```